improved text array and working on text output

+++ include/vec3.h
@@ -8,22 +8,22 @@ public:
 	vec3(double x = 0.0, double y = 0.0, double z = 0.0) : X(x), Y(y), Z(z) {}
 	double X, Y, Z;
 
-	vec3 times(double f) {
+	vec3 times(float f) {
 		return vec3(X*f, Y*f, Z*f);
 	}
-	double distance(vec3 v) {
+	float distance(vec3 v) {
 		return sqrt(sqr(v.X - X) + sqr(v.Y - Y) + sqr(v.Z - Z));
 	}
 	vec3 operator+(vec3 v) {
 		return vec3(X + v.X, Y + v.Y, Z + v.Z);
 	}
 private:
-	static double sqr(double x) { return x*x; }
+	static float sqr(float x) { return x*x; }
 };
 class vec3I {//ints
 public:
 	vec3I(int x = 0, int y = 0, int z = 0) : X(x), Y(y), Z(z) {}
-	double X, Y, Z;
+	int  X, Y, Z;
 
 	vec3I times(float f) {
 		return vec3I(X*f, Y*f, Z*f);
